add regression test for custom background images
- initialize separate test for verifying bg changes in component library
- utilize pointfreeco's SnapshotTesting for image snapshotting + diffing
- utilize snapshot heic package for reducing filesize of screenshots
- add initial background image for tests to check against
